Move useAsyncFetch into shared JS
